+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this extension will be documented in this file.
 
+## [0.8.1] - 2020-07-13
+### Fixed 
+- Flux multi relation field processing
+
+## [0.8.0] - 2020-07-13
+### Added 
+- Pages raw duplicator
+
 ## [0.7.2] - 2020-06-22
 ### Fixed 
 - Do not display translations of pages in the list of available models